requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.42.0
bottle==0.12.25
bottle-websocket==0.2.9
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
docutils==0.20.1
Eel==0.16.0
filelock==3.13.1
fsspec==2023.12.1
future==0.18.3
gevent==23.9.1
gevent-websocket==0.10.1
greenlet==3.0.1
idna==3.6
importlib-metadata==7.0.0
Jinja2==3.1.2
Kivy==2.2.1
kivy-deps.angle==0.3.3
kivy-deps.glew==0.3.1
kivy-deps.sdl2==0.6.0
Kivy-Garden==0.1.5
MarkupSafe==2.1.3
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
patool==1.15.0
pefile==2023.2.7
Pillow==10.1.0
pycparser==2.21
Pygments==2.17.2
pyinstaller==6.2.0
pyinstaller-hooks-contrib==2023.10
pyparsing==3.1.1
pypiwin32==223
pywin32==306
pywin32-ctypes==0.2.2
requests==2.31.0
sympy==1.12
torch==2.1.1
torchvision==0.16.1
typing_extensions==4.8.0
urllib3==2.1.0
whichcraft==0.6.1
zipp==3.17.0
zope.event==5.0
zope.interface==6.1
